requirements.txt
africastalking==1.2.8
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
schema==0.7.7
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
